Fix package check
There was some confusion around the package check which I think I have
cleared up with this change.

There is only one package check, and it's behavior will depend on
whether package installation is enabled or not. If package installation
is enabled, the check will succeed if the package is installed OR
available. If package installation is disabled, the check will succeed
if the package is installed.